Add release scripts for python SDK

+++ internal/build-scripts/cleanup.sh
@@ -2,7 +2,7 @@
 
 set -ex
 
-if [ ${LOCAL} = "true" ]; then
+if [ "${LOCAL}" = "true" ] && [ "${RUN_TYPE}" != "release" ]; then
   # Remove dist and build folders
   rm -rf dist
   rm -rf build
